Sun, Mar 26, 2023, 12:01 PM +02:00

+++ week4/reverse/reverse.c
@@ -30,7 +30,7 @@ int main(int argc, char *argv[])
     // TODO #3
     WAVHEADER header;
     fread(&header, sizeof(WAVHEADER), 1, infile);
-    int begining = ftell(infile);
+    long begining = ftell(infile);
 
     // Use check_format to ensure WAV format
     // TODO #4
